fix: Remove fixed width floating point types
BENCH: 21715681

+++ src/search/bench.cpp
@@ -29,7 +29,8 @@ void search::bench::run(searcher& searcher, const u32 depth) {
     const auto seconds = elapsed / 1000;
 
     std::cout << std::format("\ninfo string {} seconds", seconds) << std::endl;
-    std::cout << std::format("{} nodes {} nps", total_nodes,
-                             static_cast<usize>(static_cast<f64>(total_nodes) / seconds))
+    std::cout << std::format(
+        "{} nodes {} nps", total_nodes,
+        static_cast<u64>(static_cast<double>(total_nodes) / static_cast<double>(seconds)))
               << std::endl;
 }
